README.md
This project is an example of how to use the router (version 4.3.0) to navigate between the screens. The great advantage of using it is that the router implementation resolves all browsing in rendering and not in configuration files. It facilitates the use of code spliting.

Before running you should use.
yarn install

Project runs in Expo environment.